fix(blink.cmp): remove dedicated BlinkCmpDoc highlight
The default `BlinkCmpDoc` highlight is linked to `NormalFloat`, which is
a sensible default. However, due to overriding the default link with
only `fg` provided, the `bg` is cleared and the documentation popup is
indistinguishable from regular text.

This change is technically a no-op in terms of intended palette, because
`NormalFloat` should already use `palette.text`, and only fixes the
cleared `bg` bug.