fix(blink.cmp): set BlinkCmpDoc background to palette.highlight_low
Although the default of `NormalFloat` is a sensible value for the
documentation float, it has the same highlight as `Pmenu`, which means
that on narrow displays where the documentation window is shown below
the completion menu, it is not always clear where the completion items
end and where the documentation begins.

Setting the highlight to `palette.highlight_low` (compared to the
default of `palette.surface`) adds a visual cue to separate the
completion menu from the documentation window, without straying too much
from the defaults.